fix typeset_functions()
user @tswicegood kindly found a bug in typeset_functions that caused the
shell variable to bet set improperly on some version(s) of bash.

This change assumes that if the SHELL variable is set, it is set
properly.  This may not be a safe assumption, for instance if a user
write()s functions into a shell script under bash, and gives it a
shebang line of /bin/ksh, SHELL may still be set to /bin/bash when the
script executes, and this may fail.