Fix issue #2
Some shells, like zsh, may have options like 'noclobber' set which
disallows writing to files with '>' and '>>' if the file exists. I had
been using the 'mktemp' utility as a portable way to generate unique
temporary filenames, but mktemp also touch()es the files as part of its
operation, which conflicts with the noclobber option in zsh.

The solution I went with here was to unlink the file after mktemp
touches it, prior to redirecting output into it with '>'.

Write plumbing test for new _temp_filename_for function.

Update copyright notice.