Use external script for faster metadata lookup
Real-time introspection of the current shell environment is far too
slow to be practical.

Take advantage of the composure dir filesystem as a source-of-truth
for efficient access and parsing of metadata, to reimplement the
current glossary() and reference() functions as aliases to this
scripted lookup tool.

Why Perl5? It has the best cross-platform support. It handily beats Python for
this use-case, and oddly, also AWK, which has GNU vs non-GNU distribution
issues. It's highly optimized for this usage, and should be quite performant on
just about every Linux-like system out there.

Support color override via COMPOSURE_REFERENCE_COLOR env var.